requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
azure-core==1.31.0
azure-storage-blob==12.23.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Mail==0.10.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
MarkupSafe==3.0.1
multidict==6.1.0
openai==0.28.0
packaging==24.1
pillow==11.0.0
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.12
pytesseract==0.3.13
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.17.0
